pyproject.toml
[tool.poetry]
name = "wikipedia-rag"
version = "0.1.0"
description = ""
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
haystack-ai = "^2.14.0"
wikipedia = "^1.4.0"
langchain = "^0.3.25"
sentence-transformers = "^4.1.0"
faiss-cpu = "^1.11.0"
python-dotenv = "^1.1.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
